pyproject.toml
[tool.poetry]
name = "openlib"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.7.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
qtawesome = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

